move interface to common
